FIX: -Problème à l'inscription : des champs étaient déclarés non NULL dans la table membre alors qu'ils étaient non obligatoires. -Correction de la redirection après inscription qui prenait encore en compte le champs mail, qui n'est plus obligatoire. -Affichage d'image par défaut lorsqu'une photo de profil n'existe pas sur le serveur -Prise en compte des retours à la ligne pour l'affichage des commentaires -Sécurisation du chargement d'image dans les albums (vérification qu'il s'agit bien d'un fichier image) -Déplacement des fichiers sql dans un répertoire sql

+++ profil.php
@@ -192,7 +192,9 @@ if (isset($_GET['id']) and $_GET['id'] > 0) {
         <div class="image_album_profil">
             <?php
             foreach ($tab as $image) {
-                echo '<div class="images"><img src="Images_album/' . $image["nom"] . '" alt="photo album" title="image"/></div>';
+                if(file_exists('Images_album/' . $image["nom"])){
+                    echo '<div class="images"><img src="Images_album/' . $image["nom"] . '" alt="photo album" title="image"/></div>';
+                }
             }
             ?>
         </div>
